Initial commit.
Right now, the program loads shaders from vshad.txt and fshad.txt.
It displays the results of the shaders on a quad larger than the
output window. The only input to the shader is the time, which is
a float with units of 1s.

In the future, it might be more interesting. Loadable geometry,
more input to the shaders, etc.